change project structure : divide by functions (muscle, image, record...)
