+++ src/main/java/codeenthusiast/TrainingCenterApp/TrainingCenterAppApplication.java
@@ -6,8 +6,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class TrainingCenterAppApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(TrainingCenterAppApplication.class, args);
-	}
-
+    public static void main(String[] args) {
+        SpringApplication.run(TrainingCenterAppApplication.class, args);
+    }
 }
