--- src/main/java/codeenthusiast/TrainingCenterApp/entities/Image.java
@@ -1,51 +1,0 @@
-package codeenthusiast.TrainingCenterApp.entities;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Entity(name = "images")
-public class Image {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String filePath;
-
-    private String fileUrl;
-
-    public Image() {
-    }
-
-    public Image(String filePath, String fileUrl) {
-        this.filePath = filePath;
-        this.fileUrl = fileUrl;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getFilePath() {
-        return filePath;
-    }
-
-    public void setFilePath(String filePath) {
-        this.filePath = filePath;
-    }
-
-    public String getFileUrl() {
-        return fileUrl;
-    }
-
-    public void setFileUrl(String fileUrl) {
-        this.fileUrl = fileUrl;
-    }
-
-}
