--- src/main/java/codeenthusiast/TrainingCenterApp/entities/units/DistanceUnit.java
@@ -1,8 +1,0 @@
-package codeenthusiast.TrainingCenterApp.entities.units;
-
-public enum DistanceUnit {
-    METER,
-    KILOMETER,
-    MILE,
-    FEET
-}
