--- src/main/java/codeenthusiast/TrainingCenterApp/entities/units/RepetitionUnit.java
@@ -1,6 +1,0 @@
-package codeenthusiast.TrainingCenterApp.entities.units;
-
-public enum RepetitionUnit {
-    TIME,
-    MOTION
-}
