--- src/main/java/codeenthusiast/TrainingCenterApp/entities/units/WeightUnit.java
@@ -1,8 +1,0 @@
-package codeenthusiast.TrainingCenterApp.entities.units;
-
-public enum WeightUnit {
-    KILOGRAMS,
-    POUNDS,
-    BARS,
-    BODYWEIGHT;
-}
