+++ src/main/java/codeenthusiast/TrainingCenterApp/image/ImageRepository.java
@@ -1,0 +1,8 @@
+package codeenthusiast.TrainingCenterApp.image;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ImageRepository extends JpaRepository<Image, Long> {
+}
