+++ src/main/java/codeenthusiast/TrainingCenterApp/record/PersonalRecords.java
@@ -1,0 +1,54 @@
+package codeenthusiast.TrainingCenterApp.record;
+
+import codeenthusiast.TrainingCenterApp.record.custom.CustomRecord;
+import codeenthusiast.TrainingCenterApp.record.endurance.EnduranceRecord;
+import codeenthusiast.TrainingCenterApp.record.strength.StrengthRecord;
+
+import javax.persistence.*;
+import java.util.ArrayList;
+import java.util.List;
+
+@Entity
+@Embeddable
+public class PersonalRecords {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @OneToMany
+    private List<StrengthRecord> strengthRecords = new ArrayList<>();
+
+    @OneToMany
+    private List<EnduranceRecord> enduranceRecords = new ArrayList<>();
+
+
+    @OneToMany
+    private List<CustomRecord> customRecords = new ArrayList<>();
+
+
+    private void addStrengthRecord(StrengthRecord strengthRecord) {
+        strengthRecords.add(strengthRecord);
+    }
+
+    private void removeStrengthRecord(StrengthRecord strengthRecord) {
+        strengthRecords.remove(strengthRecord);
+    }
+
+    private void addEnduranceRecord(EnduranceRecord enduranceRecord) {
+        enduranceRecords.add(enduranceRecord);
+    }
+
+    private void removeStrengthRecord(EnduranceRecord enduranceRecord) {
+        enduranceRecords.remove(enduranceRecord);
+    }
+
+    private void addCustomRecord(CustomRecord customRecord) {
+        customRecords.add(customRecord);
+    }
+
+    private void removeCustomRecord(CustomRecord customRecord) {
+        strengthRecords.remove(customRecord);
+    }
+
+}
