+++ src/main/java/codeenthusiast/TrainingCenterApp/record/strength/StrengthRecordRepository.java
@@ -1,0 +1,16 @@
+package codeenthusiast.TrainingCenterApp.record.strength;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+
+@Repository
+public interface StrengthRecordRepository extends JpaRepository<StrengthRecord, Long> {
+
+    List<StrengthRecord> findByExerciseName(String exerciseName);
+
+    void deleteByExerciseNameAndWeightAndRepetition(String exerciseName, int weight, int repetition);
+
+    List<StrengthRecord> findByExerciseNameAndRepetition(String exerciseName, int repetition);
+}
