--- src/main/java/codeenthusiast/TrainingCenterApp/repositories/MuscleRepository.java
@@ -1,18 +1,0 @@
-package codeenthusiast.TrainingCenterApp.repositories;
-
-import codeenthusiast.TrainingCenterApp.entities.Muscle;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface MuscleRepository extends JpaRepository<Muscle, Long> {
-
-
-    Optional<Muscle> findByName(String name);
-
-    void deleteByName(String name);
-
-    boolean existsByName(String name);
-}
