+++ src/main/java/codeenthusiast/TrainingCenterApp/units/DistanceUnit.java
@@ -1,0 +1,8 @@
+package codeenthusiast.TrainingCenterApp.units;
+
+public enum DistanceUnit {
+    METER,
+    KILOMETER,
+    MILE,
+    FEET
+}
