+++ src/main/java/codeenthusiast/TrainingCenterApp/units/RepetitionUnit.java
@@ -1,0 +1,6 @@
+package codeenthusiast.TrainingCenterApp.units;
+
+public enum RepetitionUnit {
+    TIME,
+    MOTION
+}
