+++ src/main/java/codeenthusiast/TrainingCenterApp/units/WeightUnit.java
@@ -1,0 +1,8 @@
+package codeenthusiast.TrainingCenterApp.units;
+
+public enum WeightUnit {
+    KILOGRAMS,
+    POUNDS,
+    BARS,
+    BODYWEIGHT;
+}
